requirements.txt
PySide6>=6.0.0
mysql-connector-python>=8.0.0
